refactor: exchange controller package and variable refactor in external signature submit handler
